requirements/common.txt
Babel==1.3
Flask==0.10.1
Flask-Assets==0.10
Flask-Babel==0.9
Flask-Cache==0.12
Flask-Jsonpify==1.1
Flask-Login==0.2.11
Flask-Mail==0.9.1
Flask-Restless==0.17.0
Flask-SQLAlchemy==2.0
Flask-Script==2.0.5
-e git://github.com/spacedogXYZ/Flask-Store.git#egg=flask-store
Flask-WTF==0.11
Jinja2==2.7.2
Mako==1.0.1
MarkupSafe==0.18
PyYAML==3.10
SQLAlchemy==1.0.6
SQLAlchemy-Utils==0.30.12
WTForms==2.1
WTForms-Components==0.9.7
Werkzeug==0.10.4
alembic==0.7.5.post2
blinker==1.3
decorator==3.4.2
httplib2==0.8
infinity==1.3
intervals==0.3.1
itsdangerous==0.23
mimerender==0.5.5
phonenumbers==7.0.4
pygeocodio==0.5.0
pystache==0.5.3
python-dateutil==2.2
python-magic==0.4.15
python-mimeparse==0.1.4
pytz==2013.9
requests==2.2.1
six==1.10.0
speaklater==1.3
twilio==3.6.5
unittest2==0.5.1
validators==0.7
webassets==0.10.1
wsgiref==0.1.2
